GH-55 Add test helpers methods for usage in multiple tests
