+++ rosbag2/CMakeLists.txt
@@ -70,6 +70,7 @@ if(BUILD_TESTING)
   ament_add_gmock(rosbag2_write_integration_test
     test/rosbag2/rosbag2_write_integration_test.cpp
     test/rosbag2/rosbag2_test_fixture.hpp
+    test/rosbag2/test_helpers.hpp
     WORKING_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR})
   if(TARGET rosbag2_write_integration_test)
     target_link_libraries(rosbag2_write_integration_test librosbag2)
@@ -78,6 +79,7 @@ if(BUILD_TESTING)
   ament_add_gmock(rosbag2_read_integration_test
     test/rosbag2/rosbag2_read_integration_test.cpp
     test/rosbag2/rosbag2_test_fixture.hpp
+    test/rosbag2/test_helpers.hpp
     WORKING_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR})
   if(TARGET rosbag2_read_integration_test)
     target_link_libraries(rosbag2_read_integration_test librosbag2)
@@ -94,6 +96,7 @@ if(BUILD_TESTING)
 
   ament_add_gmock(rosbag2_integration_test
     test/rosbag2/rosbag2_integration_test.cpp
+    test/rosbag2/test_helpers.hpp
     WORKING_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR})
   if(TARGET rosbag2_integration_test)
     target_link_libraries(rosbag2_integration_test librosbag2)
@@ -105,6 +108,7 @@ if(BUILD_TESTING)
     src/rosbag2/raw_subscription.cpp
     src/rosbag2/raw_publisher.cpp
     src/rosbag2/rosbag2_node.cpp
+    test/rosbag2/test_helpers.hpp
     WORKING_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR})
   if(TARGET rosbag2_rosbag_node_test)
     ament_target_dependencies(rosbag2_rosbag_node_test rclcpp Poco ament_index_cpp std_msgs)
