+++ rosbag2/test/rosbag2/rosbag2_rosbag_node_test.cpp
@@ -21,8 +21,10 @@
 
 #include "rclcpp/rclcpp.hpp"
 #include "std_msgs/msg/string.hpp"
+
 #include "../../src/rosbag2/rosbag2_node.hpp"
 #include "../../src/rosbag2/typesupport_helpers.hpp"
+#include "test_helpers.hpp"
 
 using namespace ::testing;  // NOLINT
 
@@ -50,8 +52,8 @@ public:
     std::vector<std::string> messages;
     size_t counter = 0;
     auto subscription = node_->create_raw_subscription(topic_name, type,
-        [this, &counter, &messages](std::shared_ptr<rcutils_char_array_t> message) {
-          messages.push_back(deserialize_string_message(message));
+        [&counter, &messages](std::shared_ptr<rcutils_char_array_t> message) {
+          messages.push_back(test_helpers::deserialize_string_message(message));
           counter++;
         });
 
@@ -59,53 +61,6 @@ public:
       rclcpp::spin_some(node_);
     }
     return messages;
-  }
-
-  std::shared_ptr<rcutils_char_array_t> serialize_string_message(std::string message)
-  {
-    auto test_message = std::make_shared<std_msgs::msg::String>();
-    test_message->data = message;
-
-    auto rcutils_allocator = rcutils_get_default_allocator();
-    auto initial_capacity = 8u + static_cast<size_t>(test_message->data.size());
-    auto msg = new rcutils_char_array_t;
-    *msg = rcutils_get_zero_initialized_char_array();
-    auto ret = rcutils_char_array_init(msg, initial_capacity, &rcutils_allocator);
-    if (ret != RCUTILS_RET_OK) {
-      throw std::runtime_error("Error allocating resources for serialized message" +
-              std::to_string(ret));
-    }
-
-    auto serialized_message = std::shared_ptr<rcutils_char_array_t>(msg,
-        [](rcutils_char_array_t * msg) {
-          int error = rcutils_char_array_fini(msg);
-          delete msg;
-          if (error != RCUTILS_RET_OK) {
-            RCUTILS_LOG_ERROR_NAMED(
-              "rosbag2", "Leaking memory. Error: %s", rcutils_get_error_string_safe());
-          }
-        });
-
-    serialized_message->buffer_length = initial_capacity;
-
-    auto string_ts = rosbag2::get_typesupport("std_msgs/String");
-
-    auto error = rmw_serialize(test_message.get(), string_ts, serialized_message.get());
-    if (error != RMW_RET_OK) {
-      throw std::runtime_error("Something went wrong preparing the serialized message");
-    }
-
-    return serialized_message;
-  }
-
-  std::string deserialize_string_message(std::shared_ptr<rcutils_char_array_t> serialized_message)
-  {
-    char * copied = new char[serialized_message->buffer_length];
-    auto string_length = serialized_message->buffer_length - 8;
-    memcpy(copied, &serialized_message->buffer[8], string_length);
-    std::string message_content(copied);
-    delete[] copied;
-    return message_content;
   }
 
   std::shared_ptr<rosbag2::Rosbag2Node> node_;
@@ -125,7 +80,7 @@ TEST_F(RosBag2NodeFixture, publisher_and_subscriber_work)
 
   auto publisher = node_->create_raw_publisher(topic_name, type);
   for (const auto & message : test_messages) {
-    publisher->publish(serialize_string_message(message));
+    publisher->publish(test_helpers::serialize_string_message(message));
   }
 
   auto subscribed_messages = subscriber_future_.get();
